backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.38",
    "uvicorn>=0.34.0",
]

[tool.mypy]
strict = true
